Optimize custom PLP code
